Class example
In Python, if the Class is empty then we have to write 'pass' so that Python interpreter understands that it is empty. Every method has 'self' as argument, just like Extended Methods in C#. And we can have properties after defining class. They are called attributes here. Meaning we can have p.x = 10 where p is an object of a class P. Now p2.x will give error. Because x attribute is specific to object p, not p2. There is Inheritance also.